refactor #29: friend -> care 리네이밍
