+++ backend/src/main/java/backend/capstone/domain/care/carerelationshipinvite/entity/CareRelationshipInviteStatus.java
@@ -1,0 +1,8 @@
+package backend.capstone.domain.care.carerelationshipinvite.entity;
+
+public enum CareRelationshipInviteStatus {
+    PENDING,
+    ACCEPTED,
+    EXPIRED,
+    CANCELED
+}
