+++ backend/src/main/java/backend/capstone/domain/care/carerelationshipinvite/repository/CareRelationshipInviteRepository.java
@@ -1,0 +1,13 @@
+package backend.capstone.domain.care.carerelationshipinvite.repository;
+
+import backend.capstone.domain.care.carerelationshipinvite.entity.CareRelationshipInvite;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CareRelationshipInviteRepository extends
+    JpaRepository<CareRelationshipInvite, Long> {
+
+    boolean existsByInviteCode(String inviteCode);
+
+    Optional<CareRelationshipInvite> findByInviteCode(String inviteCode);
+}
