--- backend/src/main/java/backend/capstone/domain/friend/carerelationshipinvite/controller/CareRelationshipInvitePageController.java
@@ -1,13 +1,0 @@
-package backend.capstone.domain.friend.carerelationshipinvite.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class CareRelationshipInvitePageController {
-
-    @GetMapping("/care-relationship/invite")
-    public String invitePage() {
-        return "forward:/care-relationship/invite.html";
-    }
-}
