--- backend/src/main/java/backend/capstone/domain/friend/carerelationshipinvite/dto/CareRelationshipInviteLinkResponse.java
@@ -1,12 +1,0 @@
-package backend.capstone.domain.friend.carerelationshipinvite.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
-
-@Builder
-public record CareRelationshipInviteLinkResponse(
-    @Schema(example = "https://passedpath.site/care-relationship/invite?inviteCode=T5rfCFFy9j")
-    String inviteLink
-) {
-
-}
