--- backend/src/main/java/backend/capstone/domain/friend/carerelationshipinvite/entity/CareRelationshipInviteStatus.java
@@ -1,8 +1,0 @@
-package backend.capstone.domain.friend.carerelationshipinvite.entity;
-
-public enum CareRelationshipInviteStatus {
-    PENDING,
-    ACCEPTED,
-    EXPIRED,
-    CANCELED
-}
